+++ nanoCAT/ff/__init__.py
@@ -16,5 +16,4 @@ __all__ = [
     'run_match_job',
     'run_ff_anionic',
     'run_ff_cationic'
-
 ]